Now ignoring overlapping pop-ups if present (no need to edit shortcut anymore)
Added few lines of code that solve the problem when the program
opens a pop-up ad instead of clicking the slow download button.
The program now first checks if a pop-up ad is present or not.
If an ad is present the program first closes the ad and after that
clicks the download button.