Ignoring overlapping pop-ups if present (no need to edit shortcut anymore) (#1)
* Now ignoring overlapping pop-ups if present (no need to edit shortcut anymore)

Added few lines of code that solve the problem when the program
opens a pop-up ad instead of clicking the slow download button.
The program now first checks if a pop-up ad is present or not.
If an ad is present the program first closes the ad and after that
clicks the download button.

* Updated README.md

Created, in my opinion, more clear installation and usage guide so even
Skyrim fans unfamilier with programming, console and GitHub are able to
download, build and use this wonderful utility without any problems.

* Update README.md

---------

Co-authored-by: Alano Terblanche <18033717+Benehiko@users.noreply.github.com>